Fix for boost v1.76.

+++ test/ExecutionFramework.cpp
@@ -117,7 +117,7 @@ bytes ExecutionFramework::panicData(util::PanicCode _code)
 {
 	return
 		m_evmVersion.supportsReturndata() ?
-		toCompactBigEndian(selectorFromSignature32("Panic(uint256)"), 4) + encode(u256(_code)) :
+		toCompactBigEndian(selectorFromSignature32("Panic(uint256)"), 4) + encode(u256(static_cast<unsigned>(_code))) :
 		bytes();
 }
